refactor: rename folder 'game_engine' into 'model'
