+++ .gitignore
@@ -1,3 +1,4 @@
 .spyproject/
 __pycache__/
 .coverage
+**/.pytest_cache/
