+++ controller/controller.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import Optional
 
-from game_engine.utils import Point2D
+from model.utils import Point2D
 from view.gui import GUI
