+++ model/fill_grid.py
@@ -3,7 +3,7 @@ from typing import Callable, Set, Tuple
 
 from overrides.overrides import overrides
 
-from game_engine.utils import Point2D
+from .utils import Point2D
 
 
 def fill_grid_dummy(function: Callable[[Point2D], None], nbr_mines: int) -> None:
